FIX: Wait for exit event before exiting (Node)
The process exit within the tarsy command-line module was properly
waiting for all tests to complete, but not non-test asynchronous
processes. By waiting for the "exit" event on the process, all other
asyncrhonous events will also wait for completion. This enables testing
which was not even defined until an asynchronous process completed, etc.