FEAT: Add option to log browser output to specific DOM element
When testing in the browser, it is useful to be able to send log output
to a specific DOM element rather than document.body in some cases - such as
when integrating Tarsy into other browser-based applications.

Add option "logOutput" which can be set to any arbitrary DOM element to
have output sent to that element.